Tailor the PyPA template for this project
The original project template from
https://github.com/pypa/sampleproject/

Update the license and readme with my details.

Review and update the majority of setup.py to reflect the
intent of the project. Include the "canopen" dependency even
though code does not yet use it. Leave the data file installation
for a future update.

Update the tox.ini, setup.py and test.yml to require python
3.11.

Update release.yml to only try to publish on tagged pushes.

Tests:
 - Run tox to verify with all supported python versions
 - Run "python setup.py build" locally
 - Run "pip install -e .[dev]" and verify it completes successfully
   with reasonable looking output