Add some example parameter databases
Add in some example parameter databases that we can use for
various openinverter projects. This will give a base to build
the tool with.

Pretty-print the JSON so that it is easier to work with in a text
editor.

Update setup.py to include the parameter databases and remove the
sample files that are no longer required.

Tests:
 - Run tox to verify the build
 - Run "python setup.py sdist" and verify the correct files are
   present