Dump all parameters over CAN
This adds a "dumpall" command that fetches and displays all
the values on the requested node.

Communicating over CAN requires several new options to
configure the CAN bus. These all have sensible defaults and
are stored in a custom CliSetting class that can be used to
decorate new commands like "dumpall".

The print_parameters function is rolled into the listparams
command as it's very simple.

Tests:
 - Verify operation of the new command against an STM32
    Bluepill openinverter succeeds
 - Try incorrect node numbers, can channel and speeds and
    verify they don't work
 - Verify parameters against web interface
 - Check command help makes sense