Change parameter files to use floating point values
It doesn't make a lot of sense to use openinverter fixed point
values in parameter databases. This changes the format to use
floating point values instead.

The paramdb and test harness are updated to reflect the
change. Internally the system still uses fixed point values in
32-bit integers for CAN messages.

The parameter database samples are updated to reflect the new format and also to use stm32-sine v5.24.R which is the first released version of stm32-sine which will accept CAN SDO configuration of parameters.

Tests:
 - List parameters with all samples and test parameter files
 - Run updated unit tests
 - Query and configure a running stm32-sine FOC system
    "brakeregen" parameter with various correct and out of
    range values