Simplify parameter database format
This changes the parameter database JSON format to more
closely match the existing JSON format used by stm32-sine to
communicate with the web interface. The "value" field is
removed from each parameter as this is not relevant and the
unique CAN id is added to all parameters and values.

Whether a variable is a modifiable parameter or read-only
value is now indicated by the "isparam" field that is injected
into the objectdictionary.Variable class. Similarly the
"category" is injected if present in the database.

Unit tests and data are updated to reflect the changes the new
format.

The sample databases from c2000-sine and stm32-sine are
updated to the new format.

Tests:
 - Run updated unit tests
 - Run all commands in "oic" and verify correct operation and
    output