Add new "load" and "save" commands
These new commands allow an inverters parameters to be
saved over CAN to and from JSON files on disk. The format of
the JSON matches that used by the openinverter ESP8266 web
interface.

The "write" command is refactored to be a function which can
be called by the "write" command or as part of the
implementation of the "load" command.

Tests:
 - Load and save over CAN having modified params on the
   inverter
 - Load a parameter file downloaded from
  openinverter.org/parameters/ (some unsupported params are
  noted)
 - Write a single parameter and verify in web interface
- Take a param file retrieved from the web interface, load over
  CAN and then save back over CAN. Compare in a file
  comparison tool. No differences noted.
- Save over CAN and via web interface and compare with a file
  comparison tool. Some differences noted due to rounding
  errors in the web interface.