Support obtaining parameter databases from remote device
Johu has added support for requesting parameter database
JSON support for stm32 based projects. The database is stored
at Index 0x5001 and streamed in segmented form to
interested clients.

When no database is specified on the command line the
requested remote openinverter node is asked to provide the
database which is then parsed in memory prior to making any
requests.

Tests have been expanded to verify the new functionality. Also,
as stm32-sine includes the "degree" symbol in the units for
some parameters verify that Unicode is supported is
parameter databases.

Tests:
 - Run tox to verify tests on all python versions