Add support for sending commands
We support sending commands to openinverter systems by
abusing the SDO protocol a bit and setting up a fixed writable
index to accept commands. The sub-index indicates the
command and the data value a single argument to the
command.

The list of commands includes most of the terminal
commands which do not relate to parameters or can mapping.
Only load, save and reset are currently implemented by any
firmware. Unsupported commands fail gracefully with an error
message.

Tests:
 - Run the new load, save and reset commands while modifying
   and reading parameters to verify the operations requested
   took place.
 - Observe that resetting stm32-sine reports an error because
   no reponse is received. This is expected with the current
   stm32-sine new_can_module branch.
- Check the requested index for unsupported commands:
   defaults, start and stop in Wireshark
- Check the mode value for start in the Wireshark request
- Check help for the "cmd" group and each individual
   command is readable