Add a parameter logging command
This creates a way to log a specified list of parameters on a
report device. The parameters are validated before each value
is fetched and logged to the specified file in CSV format. It is
possible to specify a time step between logging operations of 1
second to 1 hour. A timestamp is provided for each row in the
CSV but can be disabled if required.

Tests:
 - Generate CSV output on the console with a range of valid and
   invalid parameters. Check "pot" value changes with throttle
  pedal sensor.
 - Generate a CSV file and load into LibreOffice. Check the data
   values can be graphed against time.
 - Disable timestamping and verify the output
 - Specify a larger than default time interval and check
   timestamps change at the expected interval