Improve exception handling and error message display
This centralises the exception handling for CAN based actions
in the can_action wrapper. An attempt has been made to
detect common CAN and CANopen SDO errors and report
them in a single line.

Errors that occur when fetching remote parameter databases
cause multiple exceptions which are not caught and
processed. No attempt has been made to fix this.

Fix a function wrapper naming error in the db_action decorator
that made understanding tracebacks hard to interpret.

Tests:
 - With a parameter database specified verify that CAN network
    down, incorrect device and unavailable CAN SDO errors are
    displayed on a single line without a complex traceback.