Add a basic CANopen SDO scanner
Use the scanning functionality in the canopen library to find
available SDO nodes on the network. To avoid overloading the
bus limit to the first 20 addresses. Going beyond this causes
odd errors.

The node ID list has to be filtered to remove internal node IDs
that canopen seems to insert in the list???

Tests:
 - Run on a network with two nodes 1 & 4 and verify they show
   up in the list
 - Turn off all devices and verify no nodes show up
 - Enable each device in turn and verify the scan finds them