Implement parameter database caching
Use the parameter database checksum and automatic
parameter database download over SDO to
uniquely(ish) identify parameter databases. This allows
them to be cached in an appropriate local directory
which speeds up the ability to manipulate parameters
greatly.

Finding a platform appropriate location for the local
cache requires the appdirs python library which is
taken as a dependency.

The way canopen is imported into paramdb.py is
refactored to use current best practice(?) and the same
way as cli.py.

No unit tests for import_cached_database() due to
laziness.

Tests:
 - With no ~/.cache/openinvter_can_tool verify that
   running the tool automatically creates the directory
 - Verify that running "oic read brakeregen" downloads
    the parameter database and writes it to
    ~/.cache/openinverter_can_tool/1-87037.json and
    the contents make sense
 - Re-run "oic read brakeregen" and verify it loads from
    the cache by checking the CAN traffic in Wireshark
 - Delete the local cached file but retain
    ~/.cache/openinverter_can_tool/1 and verify that it is
    re-downloaded without issues with "oic read
    brakeregen"
 - Sanity check other oic commands that use parameter
    database
 - Verify that manually specifying the database directly
    from the cache works with no query of the node
    checksum.