Change "can" command to "map"
This changes the high-level command for
managing can message mappings to "map"
which is an appropriate verb. Update the help
text for each sub-command to be a little clearer
as to its purpose.

The command group is changed from can to
can_map to avoid a variable clash pointed out
by flake8.

Tests:
- Review help text for the "map" command and
   each sub-command
- Verify flake8 now reports no errors or
   warnings