Remove param ID display in map add command
We don't want to display parameter IDs to the
user as they are considered dynamic and
subject to change from firmware release to
release. This change simplifies the "map add"
command to not display or allow the ID to be
specified.

Change to use modern python format strings
as the rest of the code does.

Change the display of the length parameter to
be consistent between "map list" and "map
add".

Move the ID generation from a variable to a
custom function in paramdb.py. This ought to
be moved into a custom Variable class in the
future.

Tests:
 - Add various valid and invalid mappings to test
    the different code paths. Check output
    matches previous version of the command.