Refactor OI parameters into custom variable class
This moves the openinverter specific variable handling into a new class rather than
requing various free functions to interrogate the canopen library's Variable class. With
this in place many type errors reported by Pylance go away. The handling of the OI
parameter number to index/subindex no longer litters the users of the parameter
variables.

The unit tests have been updated to test the new class but are largely a refactoring of
the old function tests.

Because the canopen.objectdictionary doesn't return a OIVariable but the base Variable
class a cast is required in some places to persuade the type checking to work. This has
no runtime effect.

Tests:
 - read and list parameters
 - map add, list and delete CAN mappings