Remove duplicate code in remote db tests
Use the new OISimulatedNode to replace lots of duplicated code in
the DatabaseImport tests.

This fixes a bug in the OISimulatedNode class which needs to
ensure that it disconnects the two network instances when the class
is cleaned up at the end of a test. This cause network failures but
only when multiple tests were run in succession.

Tests:
 - Run all unit tests
 - Ensure flake8 reports no errors