Fix database cache creation problem on Windows
On Windows the cache location is
%LOCALAPPDATA%\openinverter\openinverter_can_tool\Cache
When oic is run for the first time this path does not exist and
paramdb.import_cached_database is unable to create it.

The fix is to update the call to mkdir() in import_cached_database()
to include creation of missing parent directories.

In addition the change to use pathlib.Path objects in paramdb
broke running oic as cli.py only passed in str objects. In both calls
to paramdb create a Path object from the appropriate database
path string.

Update the paramdb unit test to include multiple directories in an
uncreated path.

Tests:
 - Run updated unit tests
 - Verify oic listparams with an explicit database works
 - Verify oic listparams without an explict database works