Fix variable type used when writing values
Fix missing type declarations in the move from
canopen.objectdictionary.Variable to paramdb.OIVariable
migration. These affect the read and write commands. Pre-
conditions are added to set_float_value() to persuade the type
checker that it is OK to evaluate param.min and param.max. We
ensure that these pre-conditions are always set when building the
database so these should never trigger.

Tests:
 - Verify that type errors in the affected functions are cleared up
 - Verify that it is possible ot write float and enum parameters
 - Verify that read and dumpall display the expected output