Split OpenInverterNode remote db into separate class
The remote database access indices and the
rest of the OpenInverterNode class are quite
distinct. Split the remote db checksum and
database string retrieval into a new
RemoteDatabaseNode class that can be used
within param_db. It could have been included
inside param_db but it complicates that
module so easier to maintain it as a separate
module.

The functions are changed from PascalCase to
snake_case to keep pylint happy.

Several type issues are fixed in the new unit
tests.

There is no coverage in test_remote_db of the
param_db() method as this is heavily exercised
in the param_db unit tests.

Tests:
 - Run split but otherwise unmodified unit tests
 - Verify no pylint, flake8 errors in new modules