Use OpenInverterNode rather than canopen.RemoteNode
Refactor the OpenInverterNode to derive from
canopen.BaseNode and behave in a similar
manner to the RemoteNode class. The
OpenInverterNode class is much simpler than
the RemoteNode class which has lots of
functionalities an openinverter device will
never expose.

The canopen.Network is passed directly to
simplify the interface. We can't really add the
node to the network in the same way as the
canopen library internal node classes but this
gains us nothing in our use case.

Tests:
 - Run all unit tests
 - Use oic read and dumpall commands to
    exercise SDO object dictionary based CAN
    network actions