Convert cli to use OpenInverterNode class
This changes all the hand rolled protocol code
to use the OpenInverterNode class to interact
with remote devices. This largely involves
deleting code and adding printing mechanisms
for the internal data structures.

To eliminate type errors the existence of
cli_settings.node is asserted in modified
commands. This should always be true.

Tests:
 - Run each command in normal operation and
    with a wide range of valid and invalid
    parameters.