Allow the creation of big-endian can map entries
Expand the "can add" command to specify the
endian-ness of the can mapping. The
OpenInverterNode interface uses a specific
endian field to specify whether big or little
endian mapping is required. The openinverter
CAN SDO protocol for addition follows listing
in using a signed 8-bit integer for the length
with negative numbers being used for big-
endian mappings.

Tests:
 - Run updated unit tests
 - Create big-endian mapping explicitly
 - Create little-endian mapping explicitly
 - Ensure that not specifying the endian-ness
    creates a little-endian mapping