Add can map import capability
This add an map_persistence.import_json_map to
mirror map_persistence.export_json_map. It is not
integrated into the main cli as further functionality is
needed in oi_node.

The CanMessage and MapEntry classes in oi_node are
extended with __eq__ and __repr__ methods to allow
them to be used in tests and in the debugger more
easily.

A reasonably comprehensive set of unit tests is
provided to exercise the normal and failure modes.

Tests:
 - Run updated unit tests