Add a way to clear all CAN map entries on a device
To allow all of the map entries to be cleared quickly
add a helper method to OpenInverterNode. This
repeatedly removes either tx.0.0 or rx.0.0 until an error
is reported. Because the device will shunt existing map
entries into the 0.0 position this will result in all CAN
message and parameter mappings being removed.

Expose the new capability with a new command "can
clear" which requires a direction to be specified.

Tests:
 - On a oi device with a complex CAN map ensure that
   each direction can be cleared independently