Add a command to import CAN message maps in JSON
Add an import capability to mirror the JSON export. By
default it clears any existing maps before importing the
TX and RX maps from the JSON.

A new method is added to OpenInverterNode to allow
whole maps to be added with one call. This is largely a
helper method so doesn't have comprehensive tests.
The tests were built from manually running the
commands which the call to add_can_map() would be
equivalent to.

Tests:
 - Run updated unit tests
 - Export a complex CAN message map
 - Extensively modify and remove map entries
 - Import the saved map with clear
 - Verify the device map matches the initial condition
 - Verify that importing again without clear doubles up
    map entries