Use negative lengths to indicate big-endian CAN mappings
Reverse ferret on the earlier work to have a separate
length and endian-ness parameters for CAN message
mappings. Stick with the openinverter/libopeninv
negative length way of representing big-endian
mappings. This is conceptually neater and consistent
between this tool and others in the openinverter
ecosystem.

Code changes are mostly mechanical. Extra validation
of zero lengths are required.

The exported map format is different. This is a breaking
change but as there has been no release with the old
format don't try to support the old format.

Tests:
 - Run updated unit tests
 - Verify that negative and positive mappings can be
    added
 - Verify zero length mappings are not allowed
 - Export and import a complex map successfully.
 - Compare exported new map against an old one and
    verify the changes are as expected (i.e. lengths of
    big-endian mappings are negative and there is no
    "endian" field in the JSON)