Allow parameters to be wildcard selected for logging
It can be a pain to log all parameters and spot values
on a device. Add wildcards of "ALL", "PARAMS" and
"VALUES" to allow each class of parameter to be
logged. The wildcard strings are upper case to avoid
clashing with real parameter names which tend to be
lower case only.

Tests:
 - Log with each of ALL, PARAMS and VALUES and verify
    output contains the expected parameters on an
    stm32-sine device
 - Log a explicit list of parameters to verify it works as
    before