Add function to export map as DBC format file
To allow integration with other CAN tools like
SavvyCAN this change includes a function that can
export a TX and RX CAN maps as a single DBC file.

The DBC file is generated using the "cantools" library
which is a new dependency.

Information on each map entry and each openinverter
parameter is translated into an appropriate "message"
and "signal" structure in the DBC file. Most information
in the map entries and parameters translates well.
Bitfields don't have an equivalent  in the DBC format
that I can see but the list of bits are provided.

The unit tests are built mostly using approval testing
due to the speculative nature of development. Each
DBC file has been opened in SavvyCAN and the
operation of each message examined for correctness
and ease of understanding.

Tests:
 - Run update unit tests
 - Where appropriate check DBC file format in
    SavvyCAN and in a text diff tool
 - Generate each TX message and verify the values in
    SavvyCAN reflect the param retrieved directly from
    the device