Change DBC export to use file path rather than File object
DBC files use a fixed CR-LF line-ending and the
cantools library has a dump_file() function that takes
account of this. This changes export_dbc_map() to take
a file path so we can use the cantools dump_file().

Each DBC generating unit test reads in the exported
file with the appropriate line-ending forced which
should allow the unit tests to pass on Windows as well
as Linux.

Tests:
 - Run all tests under tox on Linux