Fix type errors in test_paramdb
During exploratory attempts at upgrading canopen to
2.3.0 and later a few type errors were reported in
test_paramdb.py. Fix the errors by asserting we are
looking at a specific instance type of a parameter. This
should always be true for databases we create but
doesn't hold true for all canopen databases. This
change is not required for the current version of
canopen but is OK to do pre-emptively.

Tests:
 - Run existing unit tests
 - Verify errors reported with canopen forced to 2.3.0
    go away