DBC export: Assign messages to specific nodes
Rather than dump all messages in the DBC export to
the generic catchall Vector__XXX node use a specific
node names ("tx" and "rx") for messages from each
map. This makes it clearer to DBC users where
messages are intended to come from or go to the
device.

Nodes are only added to the DBC when there are
messages in that map.

Tests:
 - Run updated approval tests
 - Open the unit test generated DBC files in Savvy CAN
    and verify the messages appear under the correct
    node names