Fix #3 by scanning all 127 possible CAN open SDO nodes
Issue #3 was caused by limiting the number of nodes
to avoid hitting a lack of queue length in the Linux CAN
queue. The loop in canopen.network.scanner.search()
sends CAN frames as fast as possible. This fills the
short CAN queue that Linux sets up by default (i.e.
what everyone will use).

The fix is to rate limit to 10ms between frames by
copying the core loop from
canopen.network.scanner.search() and adding a 10ms
delay. This takes around 1.3 seconds on an idle bus
which is less than the 5 seconds we wait.

Tests:
 - Search on an idle bus with a single node
 - Search on a fairly busy bus with a single node
 - Move the node to ID 63 (the max supported by
    stm32-sine)
 - Verify the scan finds the device