Fix #4 by being more tolerant of malformed enum params
The Zombieverter parameter database described in
issue #4 has many malformed enums due to poor
punctuation. Rather than try to come up with heuristics
to try to understand this just bail out and treat them as
plain numeric values. Parameters that fail to parse are
marked with " [DB FORMAT ERROR]" in the units
description.

Add updated unit tests that exercise the various errors.

Fix the test_query_zombieverter_paramdb() which had
the wrong number of db entries as it wasn't accounting
for the "serial" parameter not being a queryable
parameter.

Tests:
 - Run updated unit tests
 - Run "oic -d zombie.json listparams" with extracted
    parameter database from
    test_query_zombieverter_paramdb() and verify that
    DB FORMAT ERROR is shown against
    "Inverter","chargemodes" and "CAN3Speed"
    parameters