Add extended frame support to JSON map persistence
Add a new field in the JSON map format to specify that
a mapped message should use an extended frame
format. The old file format is supported and all existing
can_ids are assumed to be standard rather than
extended if not specified.

Add a comprehensive set of tests to verify the import
and export.

Tests:
 - Run unit tests under tox