Bump version of JSON map format
It would be possible to import a JSON can message map with extended IDs using release 0.0.9 of the tool and corrupt the state of a device. Fix by bumping the version of the JSON and explicitly recognise both formats of the file.

The test data needs updating so that all files are version 2 (with the is_extended_frame flag) apart from the specific file used to verify version 1 format support.

Tests:
 - Run unit tests