Ensure that is_extended_frame in CanMessage is bool
Due to a faulty comparison the is_extended_frame flag
in CanMessage was actually being filled with an int
value when listing CAN mappings. This behaved
acceptably except when exported to JSON where an
odd large integer value was being stored for extended
frame mappings rather than "true". Fix by adding a
comparison and adding a type check in the unit tests.

Tests:
 - Run updated unit tests
 - With draft integration code check exported JSON
   contains a boolean value for exported maps