Implement basic CAN upgrade mechanism
Provides a new CanUpgrader class which encapsulates the upgrade.

The upgrade process consists of loading the requested firmware
image before starting a state-machine to respond to the device
being upgraded. The device is expected to be rebooted after
the CanUpgrader instance has been created to initiate the
process.

A comprehensive set of tests are included. These exercise the CAN
frames seen with existing clients as well as theoretical behaviours
observed from the implementation.

Tests:
 - Run unit tests
 - Run tox