Make CanUpgrader thread-safe
When attached to the CAN network frames are delivered in a separate thread.
The upgrader state-machine is now isolated from the caller of CanUpgrader
as copies of the state transitions are added to an internal thread-safe queue
and processed with a new run() method.

The tests require updating to the new class interface. Stopping the
state-machine mid-upgrade to test requires it to timeout. Only tests that
run to a final state (completion or failure) expect success from the run()
method.

The state-machine classes are tidied up somewhat to reflect the expected
order of execution rather than the order of implementation.

Tests:
 - Run updated unit tests