Separate internal state-machine from interface in CanUpdater
This simplifies the external interface to the CanUpdater
class by having a separate enum to describe the
possible states the upgrade process can be in. The
internal state-machine now translates to the external
interface when state transitions are added to the
thread-safe external queue.

Update the unit tests to replace extensive use of
isinstance() with simpler comparisons with the state
enum. No logical changes to any test.

Tests:
 - Run update unit tests