Add upgrade progress reporting to CanUpgrader
This adds a continuously updated progress report
through the state update callback mechanism in
CanUpdater. The progress is incremented on each
firmware page as it is confirmed successfully uploaded.
The current progress can also be accessed through a
property on the CanUpgrader class.

Existing upgrade tests are updated to add the expected
progress. An additional test is included to verify the
progress through the two page firmware upgrade.

Tests:
 - Run updated unit tests