Split CanUpgrader into two classes
Refactor the CanUpgrader so that all public methods
and properties remain available. The internal state-
machine used to process incoming CAN frames is in a
separate class which is assumed to be called on the
CAN frame callback thread. Messages are passed to
CanUpgrader over a message queue as before.

The serialno property is now added to StateUpdate to
allow it to be set and read in a thread safe manner
which it was not previously.

The tests require no changes other than to add the new field in the StateUpdate passed in callbacks.

Tests:
 - Run updated unit tests