Add a new "upgrade" command
This adds an "upgrade" command to the cli. This uses the CanUpgrader and the OpenInverterNode to fully automate firmware upgrades for operating nodes. It pulls the serial number from the device, primes the upgrade then resets the device to start the process.

Recovery with and without serial numbers is also supported.

Tests:
 - Upgrade a fully operating node
 - Ensure that serial numbers are rejected unless in
    --recover mode
 - Recover with a serial number not on the network and
    ensure the process times out
 - Recover with the correct serial number
 - Recover with no serial number
 - Review help text