Fix DBC file generation for enum parameters with an offset #8
When generating DBC files neither enum or bitfield
parameters took into account the potential that the
CAN map may have a gain and offset specified. The
resulting DBC file was then not correctly mapping
values in mapped frames.

The fix applies the gain and offset when generating the
DBC signal parameters to map. It probably only makes
sense for enum parameters to add an offset and
otherwise is fairly useless. It can be configured so it
should do the right thing.

Tests:
 - Add new approval tests having checked the output
   manually with a calculator
 - Load the approved DBC files in SavvyCAN and verify
    the output
 - Check the original example provided in Issue #8