Improve CAN upgrade unit test coverage
Due to a failure in TDD technique there are a few short
comings in code coverage of the can_upgrade module.
Add four new tests to exercise corner cases in the
protocol.

After reasoning about the UploadState.process()
method it seems impossible to ever trigger the
protocol error failure the code was hedging against.
Delete the condition.

Tests:
 - Run unit tests with coverage and ensure that
    can_upgrade now shows 100% coverage.
 - Run all tests through tox