Fix #9 by improving parsing of enums and bitfields in parameter dbs
This change tries to make the parsing of enums and
bitfields in parameter databases equivalent to the
behaviour of esp8266-web-interface. We now split the
enum/bitfield definitions by either comma or space.
This is more tolerant of poorly punctuated lists found
in real OI projects.

Update the unit tests to revise the expected behaviour
and add a new test which covers a corrupt list that
cannot be understood.

Tests:
 - Run updated unit tests
 - Verify the TorqDerate and CAN3Speed parameters on
    Zombieverter V2.20A firmware can now be read
   successfully with no "DB FORMAT ERROR" being
   displayed.