When logging output symbolic values for enums and bit-fields
To make it easier to interpret logging of bit-field and
enumeration values add a "--symbolic" flag to the log
command. This is now the default. The old behaviour
can be reinstated by supplying "--numeric".

The implementation converters the list of parameter
names to OIVariable references as this saves looking
them up repeatedly during operation.

Tests:
 - Verify ALL, VALUES and PARAMS produce the
    expected output by default and with --symbolic and
    --numeric
 - Manually specify a numerical parameter, enum and
    bit-field and verify they are output correctly
 - Specify a series of unknown and known params and
    verify the unknown ones are reported as such but the
    valid ones are used to create log entries