Add a listparam command to explain a single param
It is often useful to know the definition of a single
param especially min/max limits. Rather than listing all
params and searching back through the command
history allow a single param to be listed.

Implemented by sharing the printing of a single param
definition between listparam and listparams
commands in a similar way to "read" and "dumpall".

Use click.echo() in preference to print().

Tests:
 - Verify operation of listparams is unaffected
 - Verify listparam works with params and spot values
 - Check the help text of the listparam command