Improve test coverage of OpenInverterNode
The test coverage shows a number of minor issues with
missing unit tests.

Add some tests to cover equivalence checks for
CanMessage and MapEntry when supplied with an
incorrect type. This shouldn't happen but is best practice
to have the check and thus needs testing.

There are several instances in managing listing and
removing CAN maps where an error can occur
communicating with the remote device. Simulate the
device reporting an error and verify the response is
handled by raising an exception.

There are some missing areas of coverage:
 - Where the direction is checked to be part of an enum. It
   is hard to test this when using python typechecking.
 - SdoCommunicationErrors when removing map entries

I can't see how to test these.

Tests:
 - Run code coverage which is now 97.11% up from 94%