Sketch basic menu, toolbar and status bar of application
Create the basic UI functions for a simple workflow to
fetch parameters from a device, modify, update, save and
load locally and upgrade device firmware. Stub
implementations are provided for each action apart from
the About box.

A PyResource file is included which wraps a set of icons
from the Tango project for each function.

Tests:
 - Run on Linux and verify correct icon is displayed for
   each menu item and toolbar button as defined.