Handle and report exceptions when querying DB to fix #11
The @db_action decorator didn't catch the common
CAN related exceptions which was resulting in horrible
back-traces as described in Issue #11. The fix is to use
the same handling as the @can_action decorator.

As previously python-can's can.notifier thread spews
a lot of garbage when the interface is down. The
extent is limited to a single back-trace. Other errors
are handled cleanly when no-database is specified
and CAN is used to query the device.

Tests:
 - Test with no working CAN network
 - Test a not-present node id
 - Test a working device is unaffected