GUI:Rationalise connection management
The way CAN connections were managed was very
complex and wrong in ways that would fall apart under
some exception paths.

Simplify connection checking to use OICGui.node as
the indicator of an established connection. Add a
decorator to all UI functions that need a connection to
check for this and display an error message if not
connected.

Remove the CliSettings internal data structure from
OICGui. This pulls in a bunch of further functions from
.cli which will require further refactoring.

Keep the canopen.Network() around all of the time but
connect() and disconnect() when required. Always
ensure the connection is torn down when the
application is closed.

Tests:
 - Try all UI functions which need a CAN connection
   when not connected and connected. Verify they
   continue to function.
 - Connect and disconnect and scan for devices in
   various combinations. Verify in a debugger that only
   one reception thread is ever present.