GUI:Fix being unable to close after a failed connection
When an attempt had been made to connect when the
CAN network was down (socketcan on Linux) a
CanOperationError was thrown in the on_closing()
event handler. This prevented the application from
stopping.

The fix is to suppress CanOperationError exceptions in
on_closing(). Also, make sure that the node and
device_db are released before the netwokr is
disconnected. This is not critical but there are internal
links between nodes and the CanNetwork we are
disconnecting.

Tests:
 - Close immediately after opening
 - Attempt to connect to a down socketcan link then
   close. Reports "An error has caused receiving of
   messages to stop" on the console as expected by
   the network.check() call made by
   network.disconnect()
 - Connect, disconnect then close
 - Connect then close
 - Connect wit no valid node running then close