Make custom node classes use deterministic subscriptions
The RemoteDatabaseNode was being randomly
deleted by python when opening a tkinter file dialog.
The code in paramdb that used this assumed that it
would unsubscribe from the CAN messages as soon
as it left scope. This isn't how python works.

Now use the context manager functionality to do the
subscribe/unsubscribe to the CAN SDO messages in a
deterministic manner. This requires changes to users
to use "with".

Tests:
 - Run lightly updated unit tests
 - Run dumpall with clean cache and pre-cached
 - Press Save/Load Parameters buttons successfully in
   the GUI which was failing previously