Refactor param changing into a ParamWriter class
The new ParamWriter class can be used both by the
CLI and GUI. It has to be supplied with a logging target
which ends up in stdout for the CLI or the log window.

A set of unit tests have been created which provide
100% coverage of the new param_utils module. A bug
was identified in which incorrect bit names in a bitfield
would be silently ignored. An error is now generated
as they would be for enum parameters. No OI project
currently uses bitfield parameters I think.

Tests:
 - Load and Save Parameters in the GUI and CLI
 - Modify float and enum parameters in the GUI and CLI
    and verify they have been correctly set.