GUI: Tidy main window title and size
Adding "GUI" to the main window title is a bit
tautological.

The main window size being constrained to 800x600
causes the Parameter tab to be too small for all of the
columns on Linux and Windows (and presumably
MacOS). Removing the constrained makes the window
a little wider and accomodates all columns cleanly.

Add some missing type annotations that make code-
completion work better.

Tests:
 - Check UI layout on Linux (GNOME)
 - Manual check on Windows after push