Refactor CAN session model
This changes the CAN session to be one that requires
and explicit connection or disconnection action. The
connection is maintained in the form of an
OpenInverterNode instance on the main model.

The status of a connection is reflected in the status
bar rather than the window title.

Error reporting is centralised in the main window
which raises a QMessageBox. The controller is
responsible for catching any normal CAN errors and
return the system to a consistent state before emitting
the can_error signal.

Tests:
 - Connect to an offline CAN network and ensure an
   Network down error is reported
 - Connect to a functional CAN network and device and
   disconnect
 - Try various permutation of turning off devices and
   disabling or enabling CAN networks displays an
   appropriate error and/or recovers as it should
 - Ensure that closing the window doesn't display an
   error irrespective of the state of the system