Create a parameter model and fill it on connection
Create a parameter model derived from the
QAbstractTableModel. This stores a list of OIVariables
and their current value and can be attached later to
one or more views.

When a connection is established fetch the current
value of all parameters but exclude spot values and
put these in the new model.

Tests:
 - Connect to an inverter and verify int he debugger
   that the ParameterTableModel instance has a list of
   all of the expected parameters and sensible floating
   point values for each.