Allow viewing of parameters in a table
Update the main window to have a tabbed notebook
for the central widget. Put this somewhere we can add
future views. Also, set a sensible default size for the
window now we have something to put in it.

Create a custom ParamTableView which allows
parameters to be viewed easily and whole rows to be
selected. Set the width of the columns to be sensible
for the typical size of parameter names in OI projects.

Update the parameter model and controller to clear
parameter data from one session to the next. Without
this the data in the model was just being appended to
each time a connection was made.

Tests:
 - Connect to stm32-sine and zombieverter and view
    lists of parameters
 - Repeatedly connect to a device and ensure the list of
   parameters is updated but remains the same length
 - Review UI to make sure parameters are all visible
   with the default window size.