Create new Parameter and spot value models and views
Create new models for parameters and spot values
based on QStandardItemModel. This allows a custom
view to be built for parameters using a QTreeView. The
parameter values can be edited but all other items are
not-editable. The tree starts up not expanded which
seems fine.

The spot values are a read-only table which can be
selected but not edited.

Tests:
 - Load inverter and vcu parameter databases
 - Verify repeatedly loading the same or different
   device databases doesn't grow the table/trees
 - Verify that only values can be edited in the
   Parameter tab and nothing can be edited in the Spot
   Value tab